pyproject.toml
[project]
name = "clearwatch"
version = "0.1.0"
description = "Clear-Text & Weak-Transport Detector with Local LLM Guidance"
authors = [
    {name = "Clearwatch Team", email = "team@clearwatch.local"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["security", "network", "monitoring", "llm", "tshark"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Networking :: Monitoring",
]

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "orjson>=3.9",
    "watchfiles>=0.21",
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "requests>=2.31",
    "ipaddress2>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/clearwatch/clearwatch"
Repository = "https://github.com/clearwatch/clearwatch"
Documentation = "https://clearwatch.readthedocs.io"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
